+++ src/AppserverIo/Routlt/Results/ServletDispatcherResult.php
@@ -69,7 +69,7 @@ class ServletDispatcherResult implements ResultInterface, ServletContextAware
     /**
      * Initializes the instance with the configured result value.
      *
-     * @param string $result The result to be handled
+     * @param \AppserverIo\Routlt\Results\ResultDescriptorInterface $resultDescriptor The result descriptor instance
      */
     public function __construct(ResultDescriptorInterface $resultDescriptor)
     {