+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Features
 
 * Add interceptor functionality
+* Add PhtmlServlet for processing simple PHTML templates
 
 # Version 2.0.0-alpha3